+++ scripts/dw/index.ts
@@ -10,6 +10,7 @@ import { cmdAttach } from "./commands/attach.ts";
 import { cmdIdentify } from "./commands/identify.ts";
 import { cmdEnable } from "./commands/enable.ts";
 import { cmdDisable } from "./commands/disable.ts";
+import { cmdMakeAdmin } from "./commands/make-admin.ts";
 
 async function main(): Promise<void> {
 	const sub = process.argv[2];
@@ -48,9 +49,12 @@ async function main(): Promise<void> {
 		case "disable":
 			cmdDisable();
 			break;
+		case "make-admin":
+			cmdMakeAdmin();
+			break;
 		default:
 			fatal(
-				`unknown subcommand: ${sub} (use: setup | run | teardown | list | reset | logs | attach | identify | enable | disable)`,
+				`unknown subcommand: ${sub} (use: setup | run | teardown | list | reset | logs | attach | identify | enable | disable | make-admin)`,
 			);
 	}
 }
