+++ server/src/internal/billing/v2/actions/attach/attach.ts
@@ -77,6 +77,7 @@ export async function attach({
 		billingContext,
 		billingPlan,
 		params,
+		preview,
 	});
 
 	if (preview) {
