+++ server/src/internal/billing/v2/actions/attach/errors/handleStartDateErrors.ts
@@ -9,24 +9,18 @@ import {
 	RecaseError,
 } from "@autumn/shared";
 import { StatusCodes } from "http-status-codes";
+import { assertStripeBackdateInvoiceLineItemLimit } from "@/internal/billing/v2/utils/stripeBackdateStartDateUtils";
 
 export const handleStartDateErrors = ({
 	billingContext,
 	params,
+	preview = false,
 }: {
 	billingContext: AttachBillingContext;
 	params: AttachParamsV1;
+	preview?: boolean;
 }) => {
 	if (params.starts_at === undefined) return;
-
-	if (isPastStartDate(params.starts_at, billingContext.currentEpochMs)) {
-		throw new RecaseError({
-			message:
-				"starts_at cannot be set to a past timestamp. Use now or a future Unix timestamp in milliseconds.",
-			code: ErrCode.InvalidRequest,
-			statusCode: StatusCodes.BAD_REQUEST,
-		});
-	}
 
 	if (params.plan_schedule === "end_of_cycle") {
 		throw new RecaseError({
@@ -35,6 +29,60 @@ export const handleStartDateErrors = ({
 			code: ErrCode.InvalidRequest,
 			statusCode: StatusCodes.BAD_REQUEST,
 		});
+	}
+
+	const prices = billingContext.attachProduct.prices;
+	const isPaidRecurring =
+		!isFreeProduct({ prices }) && !isOneOffProduct({ prices });
+
+	if (isPastStartDate(params.starts_at, billingContext.currentEpochMs)) {
+		if (!isPaidRecurring) {
+			throw new RecaseError({
+				message: "Past starts_at is only supported for paid recurring plans.",
+				code: ErrCode.InvalidRequest,
+				statusCode: StatusCodes.BAD_REQUEST,
+			});
+		}
+
+		if (
+			billingContext.stripeSubscription ||
+			billingContext.stripeSubscriptionSchedule ||
+			billingContext.scheduledCustomerProduct
+		) {
+			throw new RecaseError({
+				message:
+					"Past starts_at is only supported when creating a new Stripe subscription.",
+				code: ErrCode.InvalidRequest,
+				statusCode: StatusCodes.BAD_REQUEST,
+			});
+		}
+
+		// Previews don't know whether the caller will settle via invoice (supports
+		// backdating) or Stripe Checkout (doesn't), so only block checkout on execute.
+		if (!preview && billingContext.checkoutMode === "stripe_checkout") {
+			throw new RecaseError({
+				message:
+					"Past starts_at cannot be used when Stripe Checkout is required.",
+				code: ErrCode.InvalidRequest,
+				statusCode: StatusCodes.BAD_REQUEST,
+			});
+		}
+
+		if (billingContext.trialContext?.trialEndsAt) {
+			throw new RecaseError({
+				message: "Past starts_at cannot be used together with a free trial.",
+				code: ErrCode.InvalidRequest,
+				statusCode: StatusCodes.BAD_REQUEST,
+			});
+		}
+
+		assertStripeBackdateInvoiceLineItemLimit({
+			products: [billingContext.attachProduct],
+			startsAt: params.starts_at,
+			currentEpochMs: billingContext.currentEpochMs,
+		});
+
+		return;
 	}
 
 	if (!isFutureStartDate(params.starts_at, billingContext.currentEpochMs)) {
@@ -49,9 +97,6 @@ export const handleStartDateErrors = ({
 		});
 	}
 
-	const prices = billingContext.attachProduct.prices;
-	const isPaidRecurring =
-		!isFreeProduct({ prices }) && !isOneOffProduct({ prices });
 	if (!isPaidRecurring) {
 		throw new RecaseError({
 			message: "Future starts_at is only supported for paid recurring plans.",
