+++ server/src/internal/billing/v2/actions/createSchedule/setup/setupCreateScheduleBillingContext.ts
@@ -3,6 +3,7 @@ import {
 	type CreateScheduleBillingContext,
 	type CreateScheduleParamsV0,
 	isOneOffProduct,
+	isPastStartDate,
 	isProductPaidAndRecurring,
 	type MultiAttachParamsV0,
 } from "@autumn/shared";
@@ -144,6 +145,13 @@ export const setupCreateScheduleBillingContext = async ({
 			scheduledCustomEntitlements.length > 0,
 		requestedProrationBehavior: params.billing_behavior,
 		requestedBillingCycleAnchor: params.billing_cycle_anchor,
+		billingStartsAt: immediatePhase.starts_at,
+		subscriptionBackdateStartMs: isPastStartDate(
+			immediatePhase.starts_at,
+			billingContext.currentEpochMs,
+		)
+			? immediatePhase.starts_at
+			: undefined,
 		immediatePhase,
 		futurePhases,
 		scheduledPhaseContexts,
