+++ server/tests/_groups/temp.ts
@@ -2,14 +2,33 @@ import type { TestGroup } from "./types";
 
 export const temp: TestGroup = {
 	name: "temp",
-	description: "failed retry tests",
+	description: "starts_at backdate and scheduling regression tests",
 	tier: "domain",
 	paths: [
-		"integration/billing/autumn-webhooks/billing-updated/billing-updated-attach.test.ts",
-		"integration/billing/autumn-webhooks/billing-updated/billing-updated-multi-attach.test.ts",
-		"integration/billing/autumn-webhooks/billing-updated/billing-updated-create-schedule.test.ts",
-		"integration/billing/autumn-webhooks/billing-updated/billing-updated-update-subscription.test.ts",
-		"integration/billing/autumn-webhooks/billing-updated/billing-updated-subscription-updated.test.ts",
-		"integration/billing/autumn-webhooks/billing-updated/billing-updated-subscription-deleted.test.ts",
+		"integration/billing/attach/params/start-date/starts-at-backdate.test.ts",
+		"integration/billing/attach/params/start-date/starts-at-backdate-invoice.test.ts",
+		"integration/billing/attach/params/start-date/starts-at-backdate-new-billing-subscription.test.ts",
+		"integration/billing/attach/params/start-date/starts-at-backdate-scheduled-replacement.test.ts",
+		"integration/billing/attach/params/start-date/starts-at-validation.test.ts",
+		"integration/billing/attach/params/start-date/starts-at-scheduling.test.ts",
+		"integration/billing/attach/params/start-date/starts-at-enable-plan-immediately.test.ts",
+		"integration/billing/attach/params/start-date/ends-at.test.ts",
+		"integration/billing/attach/scheduled-switch",
+		"integration/billing/attach/params/plan-schedule",
+		"integration/billing/attach/params/billing-cycle-anchor",
+		"integration/billing/attach/invoice/attach-invoice-finalized-immediate.test.ts",
+		"integration/billing/attach/invoice/attach-invoice-draft-immediate.test.ts",
+		"integration/billing/create-schedule/backdate/create-schedule-backdate.test.ts",
+		"integration/billing/create-schedule/phases/create-schedule-phases.test.ts",
+		"integration/billing/create-schedule/phases/create-schedule-phases-checkout.test.ts",
+		"integration/billing/create-schedule/phases/create-schedule-phases-replacements.test.ts",
+		"integration/billing/create-schedule/phases/create-schedule-phases-schedules.test.ts",
+		"integration/billing/create-schedule/phases/create-schedule-phases-validation.test.ts",
+		"integration/billing/create-schedule/preview/create-schedule-preview.test.ts",
+		"integration/billing/create-schedule/params/create-schedule-enable-plan-immediately.test.ts",
+		"integration/billing/create-schedule/params/create-schedule-customize.test.ts",
+		"integration/billing/create-schedule/params/create-schedule-subscription-id.test.ts",
+		"integration/billing/create-schedule/one-off-prepaid-preserve/preserve-on-schedule.test.ts",
 	],
+	maxConcurrency: 2,
 };
