+++ vite/src/components/forms/create-schedule/context/CreateScheduleFormProvider.tsx
@@ -1,7 +1,15 @@
 import type {
 	BillingPreviewResponse,
 	Feature,
+	FullCusProduct,
+	FullCustomer,
 	ProductV2,
+} from "@autumn/shared";
+import {
+	ACTIVE_STATUSES,
+	CusProductStatus,
+	isFreeProductV2,
+	isOneOffProductV2,
 } from "@autumn/shared";
 import { useStore } from "@tanstack/react-form";
 import {
@@ -9,12 +17,14 @@ import {
 	type ReactNode,
 	useCallback,
 	useContext,
+	useEffect,
 	useMemo,
 	useState,
 } from "react";
 import type { SendInvoiceSubmitParams } from "@/components/forms/shared/SendInvoiceStage";
 import { useFeaturesQuery } from "@/hooks/queries/useFeaturesQuery";
 import { useProductsQuery } from "@/hooks/queries/useProductsQuery";
+import { useCusQuery } from "@/views/customers/customer/hooks/useCusQuery";
 import {
 	type CreateScheduleForm,
 	getCreateSchedulePhaseTimingError,
@@ -47,6 +57,8 @@ interface CreateScheduleFormContextValue {
 	products: ProductV2[];
 	features: Feature[];
 	isExistingSchedule: boolean;
+	/** First phase may start in the past — only when a new Stripe subscription will be created. */
+	allowFirstPhaseBackdate: boolean;
 	isPhaseLocked: ({ phaseIndex }: { phaseIndex: number }) => boolean;
 
 	handleAddPhase: () => void;
@@ -122,6 +134,49 @@ export function CreateScheduleFormProvider({
 		[formValues.phases],
 	);
 
+	const { customer } = useCusQuery();
+	const fullCustomer = customer as FullCustomer | null;
+
+	// Mirrors the attach flow: a brand-new Stripe subscription is created when the
+	// (scoped) customer has no active/trialing subscription. Only then can the
+	// immediate phase be backdated (the server enforces the same rule).
+	const hasActiveSubscription = useMemo(() => {
+		const cusProducts = (fullCustomer?.customer_products ??
+			[]) as FullCusProduct[];
+		return cusProducts.some((cusProduct) => {
+			const activeOrTrialing =
+				ACTIVE_STATUSES.includes(cusProduct.status) ||
+				cusProduct.status === CusProductStatus.Trialing;
+			if (!activeOrTrialing) return false;
+			if (!cusProduct.subscription_ids?.length) return false;
+			const entityMatches = entityId
+				? cusProduct.entity_id === entityId ||
+					cusProduct.internal_entity_id === entityId
+				: !cusProduct.internal_entity_id;
+			return entityMatches;
+		});
+	}, [fullCustomer?.customer_products, entityId]);
+
+	const immediatePlansPaidRecurring = useMemo(() => {
+		const plans = (formValues.phases[0]?.plans ?? []).filter(
+			(plan) => plan.productId,
+		);
+		if (plans.length === 0) return false;
+		return plans.every((plan) => {
+			const product = products.find((p) => p.id === plan.productId);
+			if (!product) return false;
+			return (
+				!isFreeProductV2({ items: product.items }) &&
+				!isOneOffProductV2({ items: product.items })
+			);
+		});
+	}, [formValues.phases, products]);
+
+	const allowFirstPhaseBackdate =
+		!isExistingSchedule &&
+		!hasActiveSubscription &&
+		immediatePlansPaidRecurring;
+
 	const editingPlanValue = useMemo(() => {
 		if (!editingPlan) return null;
 		return (
@@ -161,6 +216,11 @@ export function CreateScheduleFormProvider({
 		[form.store],
 	);
 
+	const getAllowFirstPhaseBackdate = useCallback(
+		() => allowFirstPhaseBackdate,
+		[allowFirstPhaseBackdate],
+	);
+
 	const buildRequestBody = useBuildCreateScheduleRequestBody({
 		customerId,
 		entityId,
@@ -171,6 +231,7 @@ export function CreateScheduleFormProvider({
 		getBillingBehavior,
 		getResetBillingCycle,
 		getEnablePlanImmediately,
+		getAllowFirstPhaseBackdate,
 	});
 
 	const previewRequestBody = useCreateScheduleRequestBody({
@@ -182,7 +243,18 @@ export function CreateScheduleFormProvider({
 		nowMs,
 		billingBehavior: formValues.billingBehavior,
 		resetBillingCycle: formValues.resetBillingCycle,
+		allowFirstPhaseBackdate,
 	});
+
+	// When backdating is no longer allowed (e.g. plan changed to free, or scope
+	// switched to one with an active subscription), drop a stale past start so it
+	// can't leak into the request — the first phase falls back to "now".
+	useEffect(() => {
+		if (allowFirstPhaseBackdate || isExistingSchedule) return;
+		if (form.store.state.values.phases[0]?.startsAt != null) {
+			form.setFieldValue("phases[0].startsAt", null);
+		}
+	}, [allowFirstPhaseBackdate, isExistingSchedule, form]);
 
 	const phaseTimingError = useMemo(
 		() =>
@@ -219,6 +291,7 @@ export function CreateScheduleFormProvider({
 			products,
 			features,
 			isExistingSchedule,
+			allowFirstPhaseBackdate,
 			isPhaseLocked,
 			handleAddPhase,
 			handleInsertPhase,
@@ -248,6 +321,7 @@ export function CreateScheduleFormProvider({
 			products,
 			features,
 			isExistingSchedule,
+			allowFirstPhaseBackdate,
 			isPhaseLocked,
 			handleAddPhase,
 			handleInsertPhase,
